requirements.txt
flask-socketio
pyserial